Fix text width when changing viewport
A series of @media queries were causing the max width of the text to jump between 780px, up to 1140px, then down to 95% of the total width as the viewport shrunk.

So here we remove the @media queries we can control, and normalize .container to always be
max-width: 780px
width: 88vw

That way the text stays put until the viewport is small enough for the 88vw to take charge.